% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{duncan}
\alias{duncan}
\title{Duncan Reservoir}
\format{A SpatialPolygons object of Duncan Reservoir, British
Columbia.
The projection is EPSG:3153}
\usage{
duncan
}
\description{
From the BC Freshwater Atlas.
}
\keyword{datasets}

