% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eggs}
\alias{eggs}
\title{Egg Mat Data}
\format{A data.frame:
\describe{
  \item{Year}{The year of deployment as an integer.}
  \item{Dayte}{The day of the year of deployment as a Date.}
  \item{Mats}{The number of egg mats deployed as an integer.}
  \item{Days}{The number of days the mats were deployed as an integer.}
  \item{Eggs}{The number of eggs collected as an integer.}
}}
\usage{
eggs
}
\description{
For Mountain Whitefish from the Lower Duncan River, British Columbia.
}
\keyword{datasets}

