% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gsi}
\alias{gsi}
\title{GSI Data}
\format{A data.frame:
\describe{
  \item{Year}{The year of capture as an integer.}
  \item{Dayte}{The day of the year of capture as a Date.}
  \item{CaptureMethod}{The method of capture as a factor.}
  \item{ForkLength}{The fish's fork length (in mm) as an integer.}
  \item{BodyMass}{The fish's wet body mass (in g) as a numeric.}
  \item{StomachMass}{The fish's wet stomach mass (in g) as a numeric.}
  \item{GonadalMass}{The fish's wet gonadal mass (in g) as a numeric.}
  \item{Sex}{The fish's sex as a factor with two levels: Male and Female.}
  \item{GSI}{The fish's Gonadal Somatic Index as a numeric.}
}}
\usage{
gsi
}
\description{
For Mountain Whitefish from the Lower Duncan River, British Columbia.
}
\keyword{datasets}

