% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kootenay}
\alias{kootenay}
\title{Kootenay Lake}
\format{A SpatialPolygons object of Kootenay Lake, British
Columbia.
The projection is EPSG:3153}
\usage{
kootenay
}
\description{
From the BC Freshwater Atlas (manually corrected).
}
\keyword{datasets}

