% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lardeau}
\alias{lardeau}
\title{Lardeau River}
\format{A SpatialPolygons object of the Lardeau River, British
Columbia.
The projection is EPSG:3153}
\usage{
lardeau
}
\description{
From the BC Freshwater Atlas.
}
\keyword{datasets}

