% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ldr}
\alias{ldr}
\title{Lower Duncan River}
\format{A SpatialPolygons object of the Lower Duncan River, British
Columbia.
The projection is EPSG:3153}
\usage{
ldr
}
\description{
From the BC Freshwater Atlas (manually corrected).
}
\keyword{datasets}

