% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mats}
\alias{mats}
\title{Mats}
\format{A SpatialPointsDataFrame object of the egg mat sites on the Lower Duncan River, British
Columbia.
The projection is EPSG:3153}
\usage{
mats
}
\description{
Mats
}
\keyword{datasets}

