% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mwstdatr}
\alias{mwstdatr}
\alias{mwstdatr-package}
\title{Mountain Whitefish Spawn Timing Data}
\description{
\code{mwstdatr} is the R data package for the Irvine et al. manuscript
When Do Mountain Whitefish (Prosopium williamsoni) Spawn?
A Comparison of Estimates Based on Gonadosomatic Indices and Spawner and Egg Counts.
}
\details{
For a description of the available data sets type \code{data(package = "mwst")}.
}

