% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spawners}
\alias{spawners}
\title{Spawner Count Data}
\format{A data.frame:
\describe{
  \item{Year}{The observation year as an integer.}
  \item{Dayte}{The day of the year of the observation as a Date.}
  \item{Spawners}{The number of spawners observed as an integer.}
}}
\usage{
spawners
}
\description{
For Mountain Whitefish from the Lower Duncan River, British Columbia.
}
\keyword{datasets}

