% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transect}
\alias{transect}
\title{Transect}
\format{A SpatialPointsDataFrame object of the spawner survey transect start and end locations on the Lower Duncan River, British
Columbia.
The projection is EPSG:3153}
\usage{
transect
}
\description{
Transect
}
\keyword{datasets}

